requirements.txt
numpy==1.16.3
scipy==1.3.0
tensorflow-gpu==1.14.0
tensorboard==1.14.0
matplotlib==3.0.3
jupyter==1.0.0
keras==2.3.1
tifffile==2019.7.26.2
csbdeep==0.5.1
scikit-image==0.15.0
pillow==6.0.0
h5py==2.9.0